fix: PostgreSQL → Dapper C# 타입 매핑 불일치 수정
- 증상: monthly-sales, patient-ages, hospital-prescriptions, kpi 400 에러
  "A parameterless default constructor or one matching signature (...) is required"
- 원인: PostgreSQL COUNT() → bigint(Int64), ROUND()::numeric → decimal 반환
  C# 모델이 int/double로 선언되어 Dapper 역직렬화 실패
- 수정:
  - MonthlySales.PrescriptionCount: int → long
  - PatientAgeGroup.Count: int → long
  - HospitalPrescription.Count: int → long
  - KpiSummary.CurrentMonthPrescriptions: int → long
  - KpiSummary.CurrentMonthPatients: int → long
  - KpiSummary.SalesChangeRate: double → decimal
  - KpiSummary.PrescriptionChangeRate: double → decimal

+++ backend/Models/DashboardModels.cs
@@ -1,16 +1,16 @@
 namespace PharmSight.Api.Models;
 
 /// <summary>월별 매출 및 조제 건수 집계 모델</summary>
-public record MonthlySales(string Month, decimal TotalAmount, int PrescriptionCount);
+public record MonthlySales(string Month, decimal TotalAmount, long PrescriptionCount);
 
 /// <summary>약품 유형별(Rx/OTC) 매출 비중 모델</summary>
 public record DrugTypeSales(string Type, string Label, decimal Amount);
 
 /// <summary>환자 연령대 분포 모델</summary>
-public record PatientAgeGroup(string AgeGroup, int Count);
+public record PatientAgeGroup(string AgeGroup, long Count);
 
 /// <summary>의료기관별 처방전 유입 건수 모델</summary>
-public record HospitalPrescription(string HospitalName, int Count);
+public record HospitalPrescription(string HospitalName, long Count);
 
 /// <summary>도매상별 누적 지출 현황 모델</summary>
 public record WholesaleExpense(string WholesaleName, decimal Amount);
@@ -21,9 +21,9 @@ public record DrugCoverage(string Label, decimal Amount);
 /// <summary>대시보드 KPI 요약 모델</summary>
 public record KpiSummary(
     decimal CurrentMonthSales,
-    int CurrentMonthPrescriptions,
-    int CurrentMonthPatients,
+    long CurrentMonthPrescriptions,
+    long CurrentMonthPatients,
     decimal CurrentMonthOrderAmount,
-    double SalesChangeRate,
-    double PrescriptionChangeRate
+    decimal SalesChangeRate,
+    decimal PrescriptionChangeRate
 );
